WIP mid-restructure clarifying stores vs storage vs realtime concerns for conversations
